Fix error in multiple file import
When importing multiple files without selecting one first, the filepath specified will default to the directory. 
Before, the parser would assume it's a file and try to find the directory but instead would find the directory's directory, resulting in an error later if that folder didn't happen also to have the files you were importing.