Initial support for loading VtMB .mdl and .vtf files
These files are created by a custom early version of the Source engine if I understand
correctly. The files have version 2531 and have slightly different data in the .mdl files.
The textures are split into two different files (.tth/.ttz), these are currently combined
and then loaded as a .vtf file, which seems to work.

The biggest difference afaict is that the .mdl file can contain compressed vertices for the
models, this implementation does the same thing as Crowbar does more or less. Currently
only compressed vertices are supported, there's also skinned and unskinned to support.