Prevent mounted resource spam
When importing multiple models from a game at different periods of time, it will lead to a build up of discovered resources. Now when it attempts to add a discovered resource, it will check if a resource with the same name exists. If it does *and* the resource has the same path as the discovered one, then continue.

This helps speed up importing time.