Added extra condition when checking for phys UserFriendlyName as not to override the original if empty + S2 consistency
When the mesh for a phys part is being created, there's a check to see if there's a more UserFriendlyName that can be used and SourceIO will prefer that over the original (usually named after a bone). However, there is almost always a UserFriendlyName field in all vphys (atleast for Half-Life: Alyx) which causes the perfectly fine name to be overridden even if the UserFriendlyName is blank but present. This adds in a check to see if there is actually contents inside of the UserFriendlyName field.

Additionally, to replicate Source2 Modeldoc behaviour (atleast for Half-Life: Alyx), phys meshes that have neither a shape_name nor a UserFriendlyName get the name "undefined".

Tested and works in Blender 4.1.